Move classes to decoration package
